Fixed logging
