+++ packages/core/src/lib/index.ts
@@ -1,1 +1,3 @@
 export * from "./uuid";
+export * from "./logger";
+export * from "./userAgent";
