+++ packages/core/src/utils/buffer.ts
@@ -1,3 +1,5 @@
+import { getLogger } from "../lib";
+
 /**
  * Parse JSON data from a Uint8Array buffer
  */
@@ -5,7 +7,8 @@ export function parseJSONFromBuffer<T = unknown>(data: Uint8Array): T {
   try {
     return JSON.parse(new TextDecoder().decode(data)) as T;
   } catch (err) {
-    console.error("Failed to parse JSON from buffer:", err);
+    const logger = getLogger();
+    logger.error(`Failed to parse JSON from buffer: ${err}`);
     return {} as T;
   }
 }
@@ -17,7 +20,8 @@ export function parseStringFromBuffer(data: Uint8Array): string | null {
   try {
     return new TextDecoder().decode(data);
   } catch (err) {
-    console.error("Failed to parse string from buffer:", err);
+    const logger = getLogger();
+    logger.error(`Failed to parse string from buffer: ${err}`);
     return null;
   }
 }
