+++ packages/sidecar/src/main.ts
@@ -5,6 +5,7 @@ import { brotliDecompressSync, gunzipSync, inflateSync } from "node:zlib";
 import { StreamableHTTPTransport } from "@hono/mcp";
 import { type ServerType, serve } from "@hono/node-server";
 import { addEventProcessor, captureException, getTraceData, startSpan } from "@sentry/node";
+import { generateUuidv4 } from "@spotlightjs/core";
 import { Hono } from "hono";
 import { contextStorage } from "hono/context-storage";
 import { cors } from "hono/cors";
@@ -15,7 +16,6 @@ import { logIncomingEvent, logOutgoingEvent } from "./debugLogging.js";
 import { EventContainer } from "./eventContainer.js";
 import { type SidecarLogger, activateLogger, enableDebugLogging, logger } from "./logger.js";
 import { createMcpInstance } from "./mcp/index.js";
-import { generateUuidv4 } from "./messageBuffer.js";
 import { streamSSE } from "./streaming.js";
 import { parseBrowserFromUserAgent } from "./userAgent.js";
 import { type HonoEnv, type IncomingPayloadCallback, getBuffer } from "./utils.js";
